move around exceptions
